refactor clock to be a top-level module with generic type

+++ benches/sim_bench.rs
@@ -14,7 +14,7 @@ impl AdderAgent {
 }
 
 impl Agent for AdderAgent {
-    fn step(&mut self, _: &mut Option<Vec<u8>>, time: &f64, _: &mut Mailbox) -> Event {
+    fn step(&mut self, _: &mut Option<Vec<u8>>, time: &u64, _: &mut Mailbox) -> Event {
         self.sum += 1;
 
         Event::new(*time, self.id, Action::Wait)
@@ -30,7 +30,7 @@ fn run_sim(id: usize, config: Config) {
     let agent = AdderAgent::new(id);
 
     world.spawn(Box::new(agent));
-    world.schedule(0.0, id).unwrap();
+    world.schedule(0, id).unwrap();
 
     world.run().unwrap();
 }
